refactor(comply): split check.rs 2286→268 lines for CB-040 file health
check.rs grew past the 2000-line critical threshold during the coverage
bundle merge, flipping File Health (CB-040) to Fail and dropping the
composite score below the release gate.

Split into 9 sibling files via include!(), each ≤500 lines:
- check_builders_{foundation,contracts,commits,work}.rs — topical
  builder functions assembling the compliance check vector
- check_individual_{basic,cb,ci}.rs — per-check leaf functions
- check_handlers_tests_inline.rs — inline unit tests
- check_pv_enforcement_helpers_tests.rs — PV helper unit tests

Preserves visibility via module-scope inclusion; no public API change.
No unwrap/panic/expect in the extracted production code.
