+++ src/cli/handlers/comply_handlers/check_handlers/check_individual_ci.rs
@@ -1,0 +1,109 @@
+pub(crate) fn check_cargo_lock(project_path: &Path) -> ComplianceCheck {
+    debug_assert!(
+        project_path.exists(),
+        "project_path must exist: {}",
+        project_path.display()
+    );
+    if !project_path.join("Cargo.toml").exists() {
+        return ComplianceCheck {
+            name: "Cargo.lock Present".into(),
+            status: CheckStatus::Skip,
+            message: "Not a Rust project (no Cargo.toml)".into(),
+            severity: Severity::Info,
+        };
+    }
+    if project_path.join("Cargo.lock").exists() {
+        ComplianceCheck {
+            name: "Cargo.lock Present".into(),
+            status: CheckStatus::Pass,
+            message: "Cargo.lock present - reproducible builds enabled".into(),
+            severity: Severity::Info,
+        }
+    } else {
+        ComplianceCheck {
+            name: "Cargo.lock Present".into(),
+            status: CheckStatus::Fail,
+            message: "Missing Cargo.lock - run 'cargo build' to generate".into(),
+            severity: Severity::Error,
+        }
+    }
+}
+
+#[provable_contracts_macros::contract("pmat-core.yaml", equation = "path_exists")]
+pub(crate) fn check_msrv(project_path: &Path) -> ComplianceCheck {
+    debug_assert!(
+        project_path.exists(),
+        "project_path must exist: {}",
+        project_path.display()
+    );
+    let cargo_toml = project_path.join("Cargo.toml");
+    if !cargo_toml.exists() {
+        return ComplianceCheck {
+            name: "MSRV Defined".into(),
+            status: CheckStatus::Skip,
+            message: "No Cargo.toml found".into(),
+            severity: Severity::Info,
+        };
+    }
+    let content = fs::read_to_string(&cargo_toml).unwrap_or_default();
+    if content.contains("rust-version") {
+        ComplianceCheck {
+            name: "MSRV Defined".into(),
+            status: CheckStatus::Pass,
+            message: "rust-version field present in Cargo.toml".into(),
+            severity: Severity::Info,
+        }
+    } else {
+        ComplianceCheck {
+            name: "MSRV Defined".into(),
+            status: CheckStatus::Warn,
+            message: "No rust-version field - add to Cargo.toml for compatibility".into(),
+            severity: Severity::Warning,
+        }
+    }
+}
+
+#[provable_contracts_macros::contract("pmat-core.yaml", equation = "path_exists")]
+pub(crate) fn check_ci_configured(project_path: &Path) -> ComplianceCheck {
+    debug_assert!(
+        project_path.exists(),
+        "project_path must exist: {}",
+        project_path.display()
+    );
+    let github_workflows = project_path.join(".github").join("workflows");
+    if github_workflows.exists() && github_workflows.is_dir() {
+        let wf_count = fs::read_dir(&github_workflows)
+            .map(|e| e.filter_map(|e| e.ok()).count())
+            .unwrap_or(0);
+        if wf_count > 0 {
+            return ComplianceCheck {
+                name: "CI Configured".into(),
+                status: CheckStatus::Pass,
+                message: format!("{} GitHub Actions workflow(s) found", wf_count),
+                severity: Severity::Info,
+            };
+        }
+    }
+    if project_path.join(".gitlab-ci.yml").exists() {
+        return ComplianceCheck {
+            name: "CI Configured".into(),
+            status: CheckStatus::Pass,
+            message: "GitLab CI configured".into(),
+            severity: Severity::Info,
+        };
+    }
+    if project_path.join("Jenkinsfile").exists() {
+        return ComplianceCheck {
+            name: "CI Configured".into(),
+            status: CheckStatus::Pass,
+            message: "Jenkins pipeline configured".into(),
+            severity: Severity::Info,
+        };
+    }
+    ComplianceCheck {
+        name: "CI Configured".into(),
+        status: CheckStatus::Warn,
+        message: "No CI configuration found - add .github/workflows/".into(),
+        severity: Severity::Warning,
+    }
+}
